refactory dashoard organizer
