+++ utils/constant.py
@@ -4,6 +4,8 @@ EXPORTED_FILES_DIRECTORY = DATA_DIRECTORY / "save_data"
 GEOJSON_REGIONS =  DATA_DIRECTORY / "geojson" / "swedish_regions.geojson"
 
 
+
+
 GRAY_1 = "#CCCCCC"
 GRAY_2 = "#657072"
 GRAY_3 = "#4A606C"
